% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsdd.R
\name{lsdd}
\alias{lsdd}
\title{Get significance,mean and sd without exception value}
\usage{
lsdd(data, alpha = 0.05, method = "more")
}
\arguments{
\item{data}{the data you wanna calculating.Make sure that the first column is diferent type,and the second is value.}

\item{alpha}{level of significance test}

\item{method}{it can be 'more' when there are more than three replicates per experimental treatment;otherwise,it should be 'less'}
}
\value{
process data out
}
\description{
Get significance,mean and sd without exception value
}
\examples{
lsdd(data=yourdata)
}
\author{
Sidong-Li \href{mailto:2875620735@qq.com}{2875620735@qq.com}
}
